enhance(views): style note preview lock button
