+++ packages/dendron-plugin-views/src/components/DendronApp.tsx
@@ -20,7 +20,8 @@ import {
   Provider,
   setLogLevel,
 } from "@dendronhq/common-frontend";
-import { Layout } from "antd";
+import { Layout, Button } from "antd";
+import LockFilled from "@ant-design/icons/lib/icons/LockFilled";
 import _ from "lodash";
 import React from "react";
 import { useWorkspaceProps } from "../hooks";
@@ -178,7 +179,29 @@ function DendronVSCodeApp({ Component }: { Component: DendronComponent }) {
     }
   });
 
-  return <Component {...props} />;
+  const isLocked = props.ide.isLocked;
+
+  const handleLock = React.useCallback(() => {
+    postVSCodeMessage({
+      type: NoteViewMessageEnum.onUnlock,
+      data: {},
+      source: DMessageSource.webClient,
+    });
+  }, []);
+
+  return (
+    <>
+      <Component {...props} />
+      {isLocked && (
+        <Button
+          shape="circle"
+          icon={<LockFilled />}
+          onClick={handleLock}
+          style={{ position: "absolute", top: 33, right: 33 }}
+        />
+      )}
+    </>
+  );
 }
 
 export type DendronAppProps = {
