+++ packages/dendron-plugin-views/src/components/DendronNotePreview.tsx
@@ -93,8 +93,6 @@ const DendronNotePreview: DendronComponent = (props) => {
   const { useConfig } = engineHooks;
   useConfig({ opts: workspace });
 
-  const isLocked = props.ide.isLocked;
-
   logger.info({
     ctx,
     msg: "enter",
@@ -112,14 +110,6 @@ const DendronNotePreview: DendronComponent = (props) => {
   const { currentTheme: themeType } = useCurrentTheme();
   useMermaid({ config, themeType, mermaid, noteRenderedBody });
 
-  const handleLock = React.useCallback(() => {
-    postVSCodeMessage({
-      type: NoteViewMessageEnum.onUnlock,
-      data: {},
-      source: DMessageSource.webClient,
-    });
-  }, []);
-
   if (props.engine.error) {
     return (
       <div>
@@ -132,12 +122,7 @@ const DendronNotePreview: DendronComponent = (props) => {
     return <div>Loading...</div>;
   }
 
-  return (
-    <div>
-      {isLocked && <button onClick={handleLock}>Unlock</button>}
-      <DendronNote noteContent={noteRenderedBody} config={config} />;
-    </div>
-  );
+  return <DendronNote noteContent={noteRenderedBody} config={config} />;
 };
 
 export default DendronNotePreview;
